feat: A2AJ and CanLII clients for Canadian legal data
Eight morie_ingest_a2aj_* functions cover the A2AJ Canadian Legal Data
corpus: coverage, search and fetch on api.a2aj.ca, per-court Parquet
download and load from Hugging Face, citation network edges, and a table
of the courts the corpus does not carry. Eight morie_ingest_canlii_*
functions wrap CanLII's browse and citator endpoints for those gaps. The
native Parquet reader now decodes three-level LIST columns, which is how
the A2AJ citation fields are stored. Two CodeQL findings in the RLS
filter (an int product widened after the multiplication) are computed in
size_t.

+++ src/morie_dsp.cpp
@@ -108,7 +108,7 @@ List morie_dsp_rls_cpp(NumericVector x, NumericVector d,
 
         for (int r = 0; r < order; ++r) {
             double s = 0.0;
-            const double *Pr = &P[r * order];
+            const double *Pr = &P[static_cast<std::size_t>(r) * static_cast<std::size_t>(order)];
             for (int c = 0; c < order; ++c) s += Pr[c] * seg[c];
             Pseg[r] = s;
         }
@@ -119,7 +119,7 @@ List morie_dsp_rls_cpp(NumericVector x, NumericVector d,
         for (int r = 0; r < order; ++r) w[r] += kg[r] * err;
         const double inv_lam = 1.0 / lam;
         for (int r = 0; r < order; ++r) {
-            double *Pr = &P[r * order];
+            double *Pr = &P[static_cast<std::size_t>(r) * static_cast<std::size_t>(order)];
             const double kgr = kg[r];
             for (int c = 0; c < order; ++c) {
                 Pr[c] = (Pr[c] - kgr * Pseg[c]) * inv_lam;
